Delete obsolete file 'pages.slint'
This file has been copied from an official slint demo when the UI was recently redesigned.
The purpose of the file in the example is to have all import statements for pages in one single file.
We do all the include statements of pages (and the respective singletons) in app_window.slint though.
At the moment this file is not referenced and obsolete.
I will delete the file and keep the include structure in app_window.slint as is.

Change-Id: I481da76082c20b06bb7c4e59d6de113f752f5dbd
Reviewed-on: https://codereview.kdab.com/c/kdab/mecaps/+/132414
Reviewed-by: Miłosz Kosobucki <milosz.kosobucki@kdab.com>