Don't use mosquitto on Windows and fix startup
We don't have mosquitto on windows yet so when it's not available we
compile it out from the project.

Property subscribed-topics needs to be explicitly initialized due to a
bug in Slint.

We need to check for Unix and Pkg-config availability because pkg-config
may be available in some circumstances (e.g. with default installation
of Strawberry Perl).

Resolution is back to 800x600 until we fix initialization issue on
Windows.

Change-Id: Iaedbb9ac3fccf51c916034d4ca6ff8151d1cece0
Reviewed-on: https://codereview.kdab.com/c/kdab/mecaps/+/132595
Reviewed-by: Marco Thaller <marco.thaller@kdab.com>