Add unit tests for NetworkAccessManager
- adds unit tests using doctest
- adds project global CMake variable BUILD_TESTS
  to enable/disable build of (unit) tests

This implements white box tests for class NetworkAccessManager
and the related transfer classes deriving from AbstractTransferHandle.

In order to test libcurl calls, i.e.
- which libcurl functions are called
- with which parameters is a libcurl function called
- in which sequence do these calls occurr
... and not the actual libcurl implementation, a stub is used to
substitute libcurl. This stub uses the Fake Function Framework
and is implemented in tst_libcurl_stub.h

Additionally to the libcurl stub we use NetworkAccessManagerUnitTestHarness
to be able to
- invoke callbacks, that would otherwise be invoked by libcurl
- access a few private members of NetworkAccessManager to do
  white box testing

Change-Id: Ife967b3fee7d0c222e199332725c94e28a09beda
Reviewed-on: https://codereview.kdab.com/c/kdab/mecaps/+/132717
Reviewed-by: Miłosz Kosobucki <milosz.kosobucki@kdab.com>