Fix lifetime in http demo
HttpTransferHandle is now created anew for each invocation and is
deleted on transfer finish.

It's still not that good because we shouldn't delete object in a slot
connected to a signal defined in this object but this time we can
get away with it because it's the only slot connected to this signal.

It can be made safer once we have deferred deletion in KDFoundation.

Change-Id: Ifa16827be4c72ffedde211878dba76cfee9df741
Reviewed-on: https://codereview.kdab.com/c/kdab/mecaps/+/134060
Reviewed-by: Marco Thaller <marco.thaller@kdab.com>