Add unit tests for ApplicationEngine
This implements tests for class ApplicationEngine.
All tests use the Singleton structs generated by slint as
a interface to test against.
The testing scope also includes slint UI to a certain extent.

ApplicationEngineUnitTestHarness
is used to initialize each demo individually, injecting mocked
dependencies for the purpose of testing.
This allows for keeping ApplicationEngine::init() signature
as lean as possible only injecting dependency AppWindow.

FtpDownloadTransferHandleUnitTestHarness + FtpUploadTransferHandleUnitTestHarness
are used to call their progressCallbackImpl() methods, which would be called by curl.

HttpTransferHandleUnitTestHarness
is used to call it's writeCallbackImpl(), which would be called by curl.

NetworkAccessManagerMock
is used to introspect when transfers are registered / unregistered with
NetworkAccessManager from ApplicationEngine.

Plase consider that ApplicationEngine is implemented using
a singleton design pattern, thus the lifetime of the ApplicationEngine
object under test is equal to the runtime of the test executable running
all the tests below.
To 'reset the UI' we reset all UI Singleton members to their default values
manually at the beginning of every single SUBCASE.