Configure project for testing with CTest
This change includes CTest module and uses automated discovery of tests
using doctest's doctest_discover_tests script.
Using this script common set-up and tear-down logic cannot be shared by
multiple test cases anymore and is moved into individual test cases.
We make sure that every call of NetworkAccessManager::instance()
(that includes implicit calls via NetworkAccessManagerUnitTestHarness)
is preceeded by setting curl_multi_init_fake.return_val to allow for
proper initialization in ctor of singleton NetworkAccessManager.

We aim to have tests and their results be independent of other tests
and independent of the order of execution of the tests. I had previously
put in some effort to make sure this is especially the case when testing
initialization of singleton NetworkAccessManager which only initializes
once per lifetime and is tricky to be either deleted or 'reset'.
Turns out, I have not been successfull doing so in the past.
I.e. initialization test of NetworkAccessManager needs to be run before
all other tests otherwise NetworkAccessManager is already initialized and
the initialization test will fail. (It is okay to run other tests
individually and in any order without running initialization test first.)
I yet again was not successfull finding a way to have the init tests be
independent of execution order. For now I suggest to accept this
peculiarity which is additionaly documented as a comment in the test.