Add tests for MQTT integration
This adds tests for MQTT integration using FakeIt [1]
as a mocking framework that is compatible with doctest.

The tests implemented in tst_mqtt.cpp ensure that calls
to the new wrapper mosquitto_wrapper.h happen in the
right order and with correct arguments.
To do so, classes MosquittoLib and MosquittoClient are
mocked and injected into MqttLib and MqttClient for the
purpose of testing.

[1] https://github.com/eranpeer/FakeIt