Explicitly ignore ConnectionHandle when connecting to signal
At this point in time we never disconnect from OR block a connection to a signal.
Thus we can ignore the ConnectionHandle returned when connecting to a signal.

Do so by using std::ignore in every inctance we call `Signal::connect()`.
(Do not set KDBindings_ENABLE_WARN_UNUSED to FALSE/OFF globally because
we want to be warned whenever we don't use a return value elsewhere in KDBindings.)